% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{blue}
\alias{blue}
\title{Apply a Blue Filter to an Image}
\usage{
blue(rgb_vector, cut_off = 127)
}
\arguments{
\item{rgb_vector}{A numeric vector of length 3 representing the RGB values of a pixel.}

\item{cut_off}{Numeric value, the threshold to determine if the pixel should be blue (default is 127).}
}
\value{
A numeric vector of length 3, either c(0, 0, 255) for blue or c(0, 0, 0) for black.
}
\description{
This function applies a blue filter to an RGB pixel based on a grey cutoff value.
}
