% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{bw}
\alias{bw}
\title{Apply Grey Filter to an Image}
\usage{
bw(rgb_vector, cut_off = 127)
}
\arguments{
\item{rgb_vector}{A numeric vector of length 3 representing the RGB values of a pixel.}
}
\value{
A numeric value .
}
\description{
This function takes an RGB vector and transforms it to black and white
}
