% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{grey}
\alias{grey}
\title{Greyscale transformation}
\usage{
grey(rgb_vector)
}
\arguments{
\item{rgb_vector}{A numeric vector of length 3 representing the RGB values of a pixel.}
}
\value{
A numeric value representing the grey value of the pixel.
}
\description{
This function takes an RGB vector and applies a greyscale transformation
}
