% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{image_filter}
\alias{image_filter}
\title{Apply a Filter to an Image}
\usage{
image_filter(image, filter)
}
\arguments{
\item{image}{A file path to an image that will be processed.}

\item{filter}{A function that processes each RGB pixel of the image.}
}
\value{
An image object with the filter applied.
}
\description{
This function takes an image and applies a filter to each pixel.
The filter is applied by passing each pixel through the provided function.
}
