% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{red}
\alias{red}
\title{Red filter}
\usage{
red(rgb_vector, cut_off = 127)
}
\arguments{
\item{rgb_vector}{A numeric vector of length 3 representing the RGB values of a pixel.}

\item{cut_off}{Numeric value, the threshold to determine if the pixel should be green (default is 127).}
}
\value{
A numeric vector of length 3, either c(255, 0, 0) for red or c(0, 0, 0) for black.

A numeric vector of length 3, either c(0, 128, 0) for green or c(0, 0, 0) for black.
}
\description{
This function applies a red filter to an RGB pixel based on a grey cutoff value.

This function applies a green filter to an RGB pixel based on a grey cutoff value.
}
