% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{grey}
\alias{grey}
\title{Image filter
Applies a colour filter to an image
Requires the "magick" package to work}
\usage{
grey(rgb_vector)
}
\arguments{
\item{rgb_vector}{A numeric vector of length 3 representing the RGB values of a pixel.}

\item{image}{A string, the path to the file to be read. The file type can be any that is recognised by the image_read() function, eg .jpg}

\item{filter}{A function, the filter to be applied to the image. Chose between grey, bw, red, green, blue}
}
\value{
An image object

A numeric value representing the grey value of the pixel.
}
\description{
Image filter
Applies a colour filter to an image
Requires the "magick" package to work
}
\details{
The function uses the magick::image_read() function to read and process the image
Ensure the package is installed and loaded
Apply Grey Filter to an Image

This function takes an RGB vector and applies a greyscale transformation
}
