% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R, R/imagefilter.R
\docType{package}
\name{image_filter}
\alias{-package}
\alias{image_filter}
\title{Image filter}
\usage{
image_filter(image, filter)
}
\arguments{
\item{image}{A file path to an image that will be processed.}

\item{filter}{A function that processes each RGB pixel of the image.}
}
\value{
An image object with the filter applied.
}
\description{
This package provides preselected tools to apply colour transformations to images.
It allows the application of a grey transformation, as well as blue, green, red and black and white
It requires the "magick" package to work

This function takes an image and applies the selected filter to each pixel.
The filter is applied by passing each pixel through the provided function.
}
\details{
The function that should be used is image_filter(). It requires the path to the image and the name of the filter to be applied.
The accepted image types are the same as for image_read() from the magick package, i.e. .png and .jpg.
The available filters are grey, blue, green, red and bw.

The function uses the magick::image_read() function to read and process the image
Ensure the package is installed and loaded
}
