% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{image_filter}
\alias{image_filter}
\title{Apply a filter to the image}
\usage{
image_filter(image, filter)
}
\arguments{
\item{image}{A file path to an image that will be processed.}

\item{filter}{A function that processes each RGB pixel of the image.}
}
\value{
An image object with the filter applied.
}
\description{
This function takes an image and applies the selected filter to each pixel.
The filter is applied by passing each pixel through the provided function.
}
\details{
The function uses the magick::image_read() function to read and process the image
Ensure the package is installed and loaded
}
